Mosquito Spawners Fixed
The mosquito spawners now function as expected. Currently, the score is tracked in the Inspector. Level Two can be activated manually in the Game Manager through the Inspector.

To switch devices to Android compatibility, go to project settings and set the section that shows "Any IOS device" and set it to Android.
*NOTE: You will need Unity Remote installed on your device.