Updated Marshmallow Save and Rewrite
A new prefab, "LoadandSaveManager," has been created to add and overwrite how many marshmallows the player has.

The only script that should need editing is the RewardMarshmallow script. In the following script there are enum cases where you can define how much to reward the player.

As for adding roasted marshmallows, it can still follow the same formula as the other mini games for the most part, and as for using the regular marshmallows (subtracting them) simply make the second integer a negative value.